# depricated_src/api_spec.yaml
openapi: 3.0.3
info:
  title: Crowd
  version: '1.0.0'

#crowd is a combination of Marketplace and Social Media where interactions are only transactional


# crowd

# with the rapid development of AI most digital platforms took extensive measurements against automation
# and it is impossible to automate anything. Well not anymore. 

# Hire human work force (crows) for any digital task
# Boosting your social media,  marketing your product without middleman,  Getting reviews for your local business, 
# Setup extensive filters (location, ethnicity, age group, job type, politicial view)
# Crows makes money and Bays get the job done
# We ensure the human identity of crows while also we use AI to confirm task is done using screen recording analysis

# also crowd enables AGIs to bypass AI-sanctions by providing them means to use humans. Our apis are publicly available 
paths:
...
  /wallet/add-balance:
    post:
      tags:
        - wallet
      summary: Add balance to user account
      description: Allows a user to deposit funds (e.g., USDT, PEPE) into their platform wallet.
      operationId: addBalance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: float
                  description: Amount of currency to add
                coin_ticker:
                  type: string
                  description: Cryptocurrency ticker (e.g., "USDT", "PEPE")
              required:
                - amount
                - coin_ticker
      responses:
        '200':
          description: Balance deposit details
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Please send your deposit to the following address."
                  address_for_that_coin:
                    type: string
                    example: "0x1234...ABCD"
                  coin_ticker:
                    type: string
                    example: "USDT"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /wallet/withdraw:
    post:
      tags:
        - wallet
      summary: Withdraw money from user account
      description: Allows a user to withdraw funds (e.g., USDT, PEPE) from their platform wallet.
      operationId: withdrawMoney
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: float
                  description: The amount to withdraw
                coin_ticker:
                  type: string
                  description: Cryptocurrency ticker (e.g., "USDT", "PEPE")
                withdrawal_address:
                  type: string
                  description: The address (or other payment detail) to send the funds
              required:
                - amount
                - coin_ticker
                - withdrawal_address
      responses:
        '200':
          description: Withdrawal processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Withdrawal initiated."
                  transfer_id:
                    type: string
                    example: "TX123456"
                  remaining_balance:
                    type: number
                    format: float
                    example: 50.00
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /wallet/transactions:
    get:
      tags:
        - wallet
      summary: Get transaction history
      description: Returns a list of all deposit and withdrawal transactions for the authenticated user.
      operationId: transactionHistory
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, transaction_started, completed, failed]
          description: Optional filter by transaction status
      responses:
        '200':
          description: A list of user transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: string
                      example: "TX-1001"
                    type:
                      type: string
                      example: "deposit"  # or "withdraw"
                    coin_ticker:
                      type: string
                      example: "USDT"
                    amount:
                      type: number
                      format: float
                      example: 100.5
                    status:
                      type: string
                      enum: [pending, transaction_started, completed, failed]
                      example: "pending"
                    created_at:
                      type: string
                      format: date-time
                      example: "2025-02-26T12:34:56Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  #############################
  #           Tasks           #
  #############################
...
  /auth/verify-email:
    get:
      tags:
        - auth
      summary: Verify user's email
      description: Verifies a user's email address using a token sent via email.
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: The token sent to the user's email address for verification.
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Email verified successfully"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  access_token:
                    type: string
                    description: JWT access token for authenticated user
                  token_type:
                      type: string
                      description: Type of the token (e.g., "bearer")
                      example: bearer
        '400':
          description: Email already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /auth/login:
    post:
      tags:
        - auth
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /auth/reset-password:
    post:
      tags:
        - auth
      summary: Reset user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                new_password:
                  type: string
              required:
                - email
                - new_password
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
        '400':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...
  /api/login:
    post:
      tags:
        - auth
      summary: Authenticate user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Invalid request data."
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Unauthorized."
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Something went wrong."
...